Added ability to add product in prodcutMAnager site Bug fix: multiple load task points the same location. With introduction with new column in locations table now we track if product is ocupied by enother task

+++ WareHouseManagerWebApp/Service/taskLocationCoordinationService.cs
@@ -18,6 +18,20 @@ namespace WareHouseManagerWebApp.Service
             if (location != null)
             {
                 location.ItemBarcode = task.ProductBarcode;
+                location.IsOnLocation = false;
+                _context.Tasks.Add(task);
+                _context.Locations.Update(location);
+                await _context.SaveChangesAsync();
+            }
+        }
+
+        public async Task AddLoadTask(taskModel task)
+        {
+            locationModel location = await _context.Locations.FindAsync(task.LocationId);
+            if (location != null)
+            {
+                location.ItemBarcode = task.ProductBarcode;
+                location.IsOnLocation = false;
                 _context.Tasks.Add(task);
                 _context.Locations.Update(location);
                 await _context.SaveChangesAsync();
@@ -32,6 +46,21 @@ namespace WareHouseManagerWebApp.Service
                 if (location != null)
                 {
                     location.ItemBarcode = null;
+                    _context.Tasks.Remove(task);
+                    _context.Locations.Update(location);
+                    await _context.SaveChangesAsync();
+                }
+            }
+        }
+
+        public async Task DeleteLoadTask(taskModel task)
+        {
+            if (task != null)
+            {
+                locationModel location = await _context.Locations.FindAsync(task.LocationId);
+                if (location != null)
+                {
+                    location.IsOnLocation = true;
                     _context.Tasks.Remove(task);
                     _context.Locations.Update(location);
                     await _context.SaveChangesAsync();
